pyproject.toml
[tool.poetry]
name = "hello_world_api"
version = "0.1.0"
description = ""
authors = ["mehd-io <mehdi@mehd.io>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
loguru = "^0.6.0"
uvicorn = {version = "==0.14.0", extras = ["standard"]}

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
tftest = "^1.6.5"
google-auth = "^2.6.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
